Questions anywhere: use question/category idnumber #311975
Moodle 3.6 added proper idnumber fields to categories and questions
so we should use those, and undo our hack that used [ID:whatever] in the
name. This includes a DB upgrade script to update any categories with
names like that to move the 'idnumber' into the real idnumber field, and
clean up the name.